Moved detailed puzzle descriptions into google drive
