+++ puzzlesolver/puzzles/README.md
@@ -1,0 +1,1 @@
+Documentation located here: [link](https://drive.google.com/drive/folders/1chP9ddGDFfeBq9nvEcZF-QTp9J-gxHvW?usp=sharing)
