+++ puzzlesolver/puzzles/__init__.py
@@ -5,14 +5,14 @@ from ..util import PuzzleException
 # Put your dependencies here
 from .hanoi import Hanoi
 from .lightsout import LightsOut
-from .pegSolitaire.pegSolitaire import Peg
+from .pegSolitaire import Peg
 from .graphpuzzle import GraphPuzzle
 from .npuzzle import Npuzzle
-from .chairs.chairs import Chairs
-from .bishop.bishop import Bishop
-from .topspin.topspin import TopSpin
-from .hopNdrop.hopNdrop import HopNDrop
-from .rubiks.rubiks import Rubiks
+from .chairs import Chairs
+from .bishop import Bishop
+from .topspin import TopSpin
+from .hopNdrop import HopNDrop
+from .rubiks import Rubiks
 
 # Add your puzzle in the puzzleList
 puzzleList = {
