--- puzzlesolver/puzzles/chairs/README.md
@@ -1,6 +1,0 @@
-# Chair Hopping
-Made by: Mark Presten
-Spring Semester 2020
-Puzzle ID: chairs
-
-## Description
