--- puzzlesolver/puzzles/hopNdrop/README.md
@@ -1,6 +1,0 @@
-# Hop N' Drop
-Made by: Mark Presten
-Fall Semester 2020
-Puzzle ID: hopNdrop
-
-## Description
