--- puzzlesolver/puzzles/pegSolitaire/README.md
@@ -1,6 +1,0 @@
-# Peg Solitaire
-Made by: Mark Presten
-Spring Semester 2020
-Puzzle ID: pegSolitaire
-
-## Description
