--- puzzlesolver/puzzles/rubiks/README.md
@@ -1,6 +1,0 @@
-# 2x2 Rubik's Cube
-Made by: Mark Presten
-Fall Semester 2020
-Puzzle ID: rubiks
-
-## Description
